+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 ## 0.68.0 (March 28th, 2018; Visual Studio Code 1.22)
 
 * extension is build for [Visual Studio Code 1.22](https://code.visualstudio.com/updates/v1_22) now
+* using new progress api for deploy operations now
+* added `$lines` function for quick code executions
 * updated the following [npm](https://www.npmjs.com/) modules:
   * [aws-sdk](https://www.npmjs.com/package/aws-sdk) `2.215.0`
   * [vscode-helpers"](https://www.npmjs.com/package/vscode-helpers") `0.26.0`
+* code cleanups and improvements
 
 ## 0.67.1 (March 28th, 2018; bugfixes)